LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/urls/base-url/document-base-url-window-initiator-is-not-opener.https.window-expected.txt


FAIL window.open() gets base url from initiator not opener. assert_equals: expected "https://web-platform.test:9443/html/infrastructure/urls/base-url/document-base-url-window-initiator-is-not-opener.https.window.html" but got "https://example.com/"
